Set type of max index to int64_t in Morton test.

+++ src/core/unit_tests/Morton_test.cpp
@@ -31,7 +31,7 @@
 BOOST_AUTO_TEST_CASE(Morton_Index_Check) {
   const int dim = 3;
   const int lvl = 20;
-  const int n_idx = 1 << (dim * lvl);
+  const int64_t n_idx = 1 << (dim * lvl);
   std::array<int64_t, 3> tmp;
   for (int64_t idx = 0; idx < n_idx; ++idx) {
     tmp = Utils::morton_idx_to_coords(idx);
